vfs_virusfilter: Fix CID 1428740 Macro compares unsigned to 0
vsnprintf returns "int" and not "size_t"


BUG: https://bugzilla.samba.org/show_bug.cgi?id=13343

(cherry picked from commit 734404bbe911fd4aa6565b2a2aaecab4fbbf3c45)

+++ source3/modules/vfs_virusfilter_utils.c
@@ -417,7 +417,7 @@ bool virusfilter_io_vwritefl(
 	const char *data_fmt, va_list ap)
 {
 	char data[VIRUSFILTER_IO_BUFFER_SIZE + VIRUSFILTER_IO_EOL_SIZE];
-	size_t data_size;
+	int data_size;
 
 	data_size = vsnprintf(data, VIRUSFILTER_IO_BUFFER_SIZE, data_fmt, ap);
